--- app/components/provider_interface/status_box_components/rejected_component.html.erb
@@ -1,15 +1,0 @@
-<h2 class="govuk-heading-m govuk-!-font-size-27">Rejection details</h2>
-
-<% if @application_choice.structured_rejection_reasons.present? %>
-  <p class="govuk-body">
-    This application was rejected on <%= @application_choice.rejected_at.to_s(:govuk_date) %>. The following feedback was sent to the candidate.
-  </p>
-  <div class="govuk-inset-text">
-    <%= render ReasonsForRejectionComponent.new(
-      application_choice: @application_choice,
-      reasons_for_rejection: ReasonsForRejection.new(@application_choice.structured_rejection_reasons),
-    ) %>
-  </div>
-<% else %>
-  <%= render SummaryListComponent.new(rows: rejected_rows) %>
-<% end %>
