+++ app/views/provider_interface/application_choices/show.html.erb
@@ -2,7 +2,7 @@
 
 <%= render ProviderInterface::ApplicationChoiceHeaderComponent.new(application_choice: @application_choice, provider_can_respond: @provider_can_respond) %>
 
-<%= render ProviderInterface::StatusBoxComponent.new(application_choice: @application_choice, options: @status_box_options) unless @offer_present %>
+<%= render ProviderInterface::StatusBoxComponent.new(application_choice: @application_choice, options: @status_box_options) unless @offer_present || @application_choice.rejected? %>
 
 <section class="app-section">
   <h2 class="govuk-heading-m govuk-!-font-size-27">Application details</h2>
