+++ spec/system/provider_interface/reject_an_application_spec.rb
@@ -180,7 +180,7 @@ RSpec.describe 'Reject an application' do
   def then_i_can_see_the_reasons_why_the_application_was_rejected
     expect(page).to have_content('Application rejected')
 
-    expect(page).to have_content('Rejection details')
+    expect(page).to have_content('Feedback')
 
     expect(page).to have_content('The following feedback was sent to the candidate.')
     expect(page).to have_content('Something you did')
